refactor: use Grid domain object and remove game concept
- change the game domain object (DO) to grid DO
- move win and draw logic from DO to evaluate use case
