+++ lib/create_new_grid.rb
@@ -1,0 +1,11 @@
+require 'grid'
+
+class CreateNewGrid
+  def initialize(grid_gateway)
+    @grid_gateway = grid_gateway
+  end
+
+  def execute
+    @grid_gateway.save(Grid.new)
+  end
+end
