--- lib/evaluate_game.rb
@@ -1,9 +1,0 @@
-class EvaluateGame
-  def initialize(game_gateway)
-    @game_gateway = game_gateway
-  end
-
-  def execute
-    @game_gateway.saved_game.win
-  end
-end
