--- lib/start_new_game.rb
@@ -1,17 +1,0 @@
-require 'game'
-
-class StartNewGame
-  def initialize(game_gateway)
-    @game_gateway = game_gateway
-  end
-
-  def execute
-    @game_gateway.save(
-      Game.new([
-        [nil, nil, nil],
-        [nil, nil, nil],
-        [nil, nil, nil]
-      ])
-    )
-  end
-end
