--- spec/test_doubles/game_gateway_fake.rb
@@ -1,7 +1,0 @@
-class GameGatewayFake
-  attr_accessor :saved_game
-
-  def save(game)
-    @saved_game = game
-  end
-end
