--- spec/test_doubles/game_gateway_spy.rb
@@ -1,7 +1,0 @@
-class GameGatewaySpy
-  attr_reader :saved_game
-
-  def save(game)
-    @saved_game = game
-  end
-end
