--- spec/test_doubles/game_gateway_stub.rb
@@ -1,3 +1,0 @@
-class GameGatewayStub
-  attr_accessor :saved_game
-end
