+++ spec/test_doubles/grid_gateway_fake.rb
@@ -1,0 +1,7 @@
+class GridGatewayFake
+  attr_accessor :saved_grid
+
+  def save(grid)
+    @saved_grid = grid
+  end
+end
