+++ spec/test_doubles/grid_gateway_spy.rb
@@ -1,0 +1,7 @@
+class GridGatewaySpy
+  attr_reader :saved_grid
+
+  def save(grid)
+    @saved_grid = grid
+  end
+end
