+++ spec/test_doubles/grid_gateway_stub.rb
@@ -1,0 +1,3 @@
+class GridGatewayStub
+  attr_accessor :saved_grid
+end
