+++ spec/unit/create_new_grid_spec.rb
@@ -1,0 +1,19 @@
+require_relative '../test_doubles/grid_gateway_spy'
+require_relative '../../lib/create_new_grid'
+
+describe CreateNewGrid do
+  it 'saves a game' do
+    grid_gateway = GridGatewaySpy.new
+    start_new_game = CreateNewGrid.new(grid_gateway)
+
+    start_new_game.execute
+
+    expect(grid_gateway.saved_grid.state).to eq(
+      [
+        [nil, nil, nil],
+        [nil, nil, nil],
+        [nil, nil, nil]
+      ]
+    )
+  end
+end
